// src/services/api/generated/models/relay.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ModelTypeEnum } from './model-type-enum';
// May contain unused imports in some cases
// @ts-ignore
import type { RelayStatusEnum } from './relay-status-enum';

export interface Relay {
    'id': string;
    /**
     * Unique name for this relay
     */
    'relay_name': string;
    /**
     * Type of relay  * `POWER_EXTENSION` - Power Extension * `ETH_008_A` - Ethernet 008 Type A * `ETH_008_B` - Ethernet 008 Type B * `ETH_016` - Ethernet 016 * `CUSTOM` - Custom
     */
    'model_type': ModelTypeEnum;
    /**
     * Relay status  * `ACTIVE` - Active * `INACTIVE` - Inactive * `MAINTENANCE` - Maintenance * `FAULT` - Fault
     */
    'status'?: RelayStatusEnum;
    /**
     * Physical location (e.g., Rack A, Shelf 2)
     */
    'location'?: string;
    /**
     * IP address of the relay
     */
    'ip_address': string;
    /**
     * MAC address in format XX:XX:XX:XX:XX:XX
     */
    'mac_address': string;
    /**
     * Number of ports on this relay
     */
    'port_count'?: number;
    'created_at': string;
    'updated_at': string;
    /**
     * Last health check timestamp
     */
    'last_checked_at': string | null;
    'is_healthy': string;
}