// src/services/api/generated/models/relay-status-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `ACTIVE` - Active * `INACTIVE` - Inactive * `MAINTENANCE` - Maintenance * `FAULT` - Fault
 */

export const RelayStatusEnum = {
    ACTIVE: 'ACTIVE',
    INACTIVE: 'INACTIVE',
    MAINTENANCE: 'MAINTENANCE',
    FAULT: 'FAULT'
} as const;

export type RelayStatusEnum = typeof RelayStatusEnum[keyof typeof RelayStatusEnum];